컴포넌트 스타일 분리 작업 [resolve #5]
- AButton Component
- Animate Component
- DropdownMenu Component
- DropdownMenuItem Component
- Header Component
- Main Component
