+++ src/components/Animate/component.js
@@ -1,18 +1,13 @@
-import styled from "styled-components";
-import cssTheme from "../../styles/css";
 import { cloneElement } from "react";
+import Styled from "./style";
 import useAnime from "../../hooks/useAnime";
-
-const Wrapper = styled.div`
-  ${cssTheme.slideAnime()};
-`;
 
 const Component = ({ children }) => {
   const [anime, setRef] = useAnime(false);
   return (
-    <Wrapper>
+    <Styled.Div>
       {cloneElement(children, { ref: setRef, anime: anime ? 1 : 0 })}
-    </Wrapper>
+    </Styled.Div>
   );
 };
 export default Component;
