+++ src/components/Animate/style.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+import cssTheme from "../../styles/css";
+
+const Div = styled.div`
+  ${cssTheme.slideAnime()};
+`;
+
+const Styled = { Div };
+export default Styled;
