+++ src/components/DropdownMenu/DropdownItem/style.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+import cssTheme from "../../../styles/css";
+
+const MenuItem = styled.div`
+  ${cssTheme.flexBox("row")}
+`;
+
+const Styled = { MenuItem };
+export default Styled;
