+++ src/components/DropdownMenu/component.js
@@ -1,24 +1,7 @@
-import styled from "styled-components";
-import cssTheme from "../../styles/css";
 import { useLayoutEffect, useRef, useState, useCallback } from "react";
+import Styled from "./style";
 import DropdownItem from "./DropdownItem";
 import useOutsideAlerter from "../../hooks/useOutsideAlerter";
-
-const Wrapper = styled.div`
-  ${cssTheme.flexBox("column", "flex-start")};
-  display: ${(props) => (props.show ? "flex" : "none")};
-  position: absolute;
-  background-color: #fff;
-  z-index: 1;
-  border-radius: 8px;
-  width: 140px;
-  height: 110px;
-  top: calc(100% + 0.25rem);
-  left: ${(props) => props.pos.x}px;
-  padding: 0.5rem;
-  box-shadow: 0 2px 5px 0 rgba(0, 0, 0, 0.1);
-  gap: 1em;
-`;
 
 const menuList = [
   {
@@ -64,9 +47,9 @@ const Component = ({ show, setShow, ...rest }) => {
     </DropdownItem>
   ));
   return (
-    <Wrapper ref={wrapperRef} pos={menuPos} show={show} {...rest}>
+    <Styled.DropDownMenu ref={wrapperRef} pos={menuPos} show={show} {...rest}>
       {items}
-    </Wrapper>
+    </Styled.DropDownMenu>
   );
 };
