+++ src/components/Header/component.js
@@ -1,37 +1,9 @@
-import styled from "styled-components";
-import cssTheme from "../../styles/css";
 import { useState } from "react";
+import Styled from "./style";
 import Voucher from "./Voucher";
 import ArrowIcon from "../ArrowIcon";
 import DropdownMenu from "../DropdownMenu";
 import AButton from "../AButton";
-
-const Wrapper = styled.header`
-  background: #fff;
-  position: fixed;
-  left: 0;
-  right: 0;
-  top: 0;
-  width: 100%;
-  height: 68px;
-  z-index: 9999;
-`;
-
-const Content = styled.div`
-  ${cssTheme.flexBox("row", "center", "space-between")};
-  margin-left: auto;
-  margin-right: auto;
-  min-width: 1024px;
-  max-width: 1164px;
-  height: 100%;
-  width: 100%;
-`;
-
-const Section = styled.div`
-  ${cssTheme.flexBox("row", "center")};
-  position: relative;
-  gap: 8px;
-`;
 
 const Component = () => {
   const [show, setShow] = useState(false);
@@ -39,9 +11,9 @@ const Component = () => {
     setShow(!show);
   };
   return (
-    <Wrapper>
-      <Content>
-        <Section>
+    <Styled.Header>
+      <Styled.Content>
+        <Styled.Area>
           <AButton href="/" notTab>
             lemonbase
           </AButton>
@@ -57,16 +29,16 @@ const Component = () => {
             비대면 바우처
             <Voucher />
           </AButton>
-        </Section>
-        <Section>
+        </Styled.Area>
+        <Styled.Area>
           <AButton href="/">로그인</AButton>
           <AButton blue="blue" href="https://lemonbase.com/free-trial/">
             무료체험 신청
           </AButton>
-        </Section>
-      </Content>
+        </Styled.Area>
+      </Styled.Content>
       <DropdownMenu show={show} setShow={setShow} />
-    </Wrapper>
+    </Styled.Header>
   );
 };
